move api calls into a directory, and make a test
